refactor!: move Display impl for Literal to codegen
